+++ torchmetrics/classification/f_beta.py
@@ -274,6 +274,7 @@ class F1(FBeta):
     """
 
     is_differentiable = False
+    higher_is_better = True
 
     def __init__(
         self,