+++ torchmetrics/regression/symmetric_mean_absolute_percentage_error.py
@@ -55,6 +55,7 @@ class SymmetricMeanAbsolutePercentageError(Metric):
         tensor(0.2290)
     """
     is_differentiable = True
+    higher_is_better = False
     sum_abs_per_error: Tensor
     total: Tensor